--- DemoApp/Properties/AssemblyInfo.cs
@@ -1,14 +1,0 @@
-﻿// This code is distributed under MIT license. Copyright (c) 2013 George Mamaladze
-// See license.txt or http://opensource.org/licenses/mit-license.php
-using System.Reflection;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyTitle("Gma.DataStructures.StringSearch.DemoApp")]
-[assembly: AssemblyDescription("An Application to demonstrate .NET Data Structures for Substring Search, Auto-completion and Intelli-sense.")]
-[assembly: AssemblyCompany("George Mamaladze")]
-[assembly: AssemblyProduct("Gma.DataStructures.StringSearch.DemoApp")]
-[assembly: AssemblyCopyright("Copyright © George Mamaladze 2013")]
-[assembly: ComVisible(false)]
-[assembly: Guid("dd3d4d0d-0bc4-4a0c-a8dd-76e70b5e9ee5")]
-[assembly: AssemblyVersion("1.0.1.*")]
-
