--- DemoApp2/Program.cs
@@ -1,14 +1,0 @@
-namespace DemoApp2 {
-    internal static class Program {
-        /// <summary>
-        ///  The main entry point for the application.
-        /// </summary>
-        [STAThread]
-        static void Main() {
-            // To customize application configuration such as set high DPI settings or default font,
-            // see https://aka.ms/applicationconfiguration.
-            ApplicationConfiguration.Initialize();
-            Application.Run(new MainForm());
-        }
-    }
-}
