--- DemoApp2/texts/QuickBrownFox.txt
@@ -1,1 +1,0 @@
-The quick brown fox jumps over the lazy dog
