+++ SampleConsoleApp/Program.cs
@@ -1,10 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.IO;
-using System.Linq;
+﻿using System.Diagnostics;
 using System.Text;
-using Gma.DataStructures.StringSearch;
 
 namespace Gma.DataStructures.StringSearch.SampleConsoleApp
 {
