+++ TrieNet.sln
@@ -12,13 +12,13 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Demo", "Demo", "{4186BBA3-FBD8-408B-AF5A-2848A4073DD1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SampleConsoleApp", "SampleConsoleApp\SampleConsoleApp.csproj", "{AA86F904-90E5-407F-87B6-688ED8A2FF85}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TrieNet", "TrieNet\TrieNet.csproj", "{1A60C4E2-8E02-48C4-BAFA-8EA0FE1BAE5E}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DemoApp", "DemoApp2\DemoApp.csproj", "{3B82BB93-FD1B-48E9-936A-FECE89810B16}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TrieNet.Test", "TrieNet.Test\TrieNet.Test.csproj", "{4EDA3CFF-EA4A-4754-9A95-750CCAF3CAB1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TrieNet.Test", "TrieNet.Test2\TrieNet.Test.csproj", "{C818EAA2-F29A-4F67-A572-8BC24683F6A4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DemoApp", "DemoApp\DemoApp.csproj", "{8B6BEE0F-F9A5-4E8F-B2DF-56CF504AA272}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SampleConsoleApp", "SampleConsoleApp\SampleConsoleApp.csproj", "{F77D3A11-DABB-4C35-B388-73E6F50628AC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -28,14 +28,6 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{AA86F904-90E5-407F-87B6-688ED8A2FF85}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AA86F904-90E5-407F-87B6-688ED8A2FF85}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AA86F904-90E5-407F-87B6-688ED8A2FF85}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{AA86F904-90E5-407F-87B6-688ED8A2FF85}.Debug|x86.Build.0 = Debug|Any CPU
-		{AA86F904-90E5-407F-87B6-688ED8A2FF85}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AA86F904-90E5-407F-87B6-688ED8A2FF85}.Release|Any CPU.Build.0 = Release|Any CPU
-		{AA86F904-90E5-407F-87B6-688ED8A2FF85}.Release|x86.ActiveCfg = Release|Any CPU
-		{AA86F904-90E5-407F-87B6-688ED8A2FF85}.Release|x86.Build.0 = Release|Any CPU
 		{1A60C4E2-8E02-48C4-BAFA-8EA0FE1BAE5E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{1A60C4E2-8E02-48C4-BAFA-8EA0FE1BAE5E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1A60C4E2-8E02-48C4-BAFA-8EA0FE1BAE5E}.Debug|x86.ActiveCfg = Debug|Any CPU
@@ -44,29 +36,37 @@ Global
 		{1A60C4E2-8E02-48C4-BAFA-8EA0FE1BAE5E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{1A60C4E2-8E02-48C4-BAFA-8EA0FE1BAE5E}.Release|x86.ActiveCfg = Release|Any CPU
 		{1A60C4E2-8E02-48C4-BAFA-8EA0FE1BAE5E}.Release|x86.Build.0 = Release|Any CPU
-		{3B82BB93-FD1B-48E9-936A-FECE89810B16}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3B82BB93-FD1B-48E9-936A-FECE89810B16}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3B82BB93-FD1B-48E9-936A-FECE89810B16}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{3B82BB93-FD1B-48E9-936A-FECE89810B16}.Debug|x86.Build.0 = Debug|Any CPU
-		{3B82BB93-FD1B-48E9-936A-FECE89810B16}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3B82BB93-FD1B-48E9-936A-FECE89810B16}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3B82BB93-FD1B-48E9-936A-FECE89810B16}.Release|x86.ActiveCfg = Release|Any CPU
-		{3B82BB93-FD1B-48E9-936A-FECE89810B16}.Release|x86.Build.0 = Release|Any CPU
-		{C818EAA2-F29A-4F67-A572-8BC24683F6A4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C818EAA2-F29A-4F67-A572-8BC24683F6A4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C818EAA2-F29A-4F67-A572-8BC24683F6A4}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{C818EAA2-F29A-4F67-A572-8BC24683F6A4}.Debug|x86.Build.0 = Debug|Any CPU
-		{C818EAA2-F29A-4F67-A572-8BC24683F6A4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C818EAA2-F29A-4F67-A572-8BC24683F6A4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C818EAA2-F29A-4F67-A572-8BC24683F6A4}.Release|x86.ActiveCfg = Release|Any CPU
-		{C818EAA2-F29A-4F67-A572-8BC24683F6A4}.Release|x86.Build.0 = Release|Any CPU
+		{4EDA3CFF-EA4A-4754-9A95-750CCAF3CAB1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4EDA3CFF-EA4A-4754-9A95-750CCAF3CAB1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4EDA3CFF-EA4A-4754-9A95-750CCAF3CAB1}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{4EDA3CFF-EA4A-4754-9A95-750CCAF3CAB1}.Debug|x86.Build.0 = Debug|Any CPU
+		{4EDA3CFF-EA4A-4754-9A95-750CCAF3CAB1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4EDA3CFF-EA4A-4754-9A95-750CCAF3CAB1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4EDA3CFF-EA4A-4754-9A95-750CCAF3CAB1}.Release|x86.ActiveCfg = Release|Any CPU
+		{4EDA3CFF-EA4A-4754-9A95-750CCAF3CAB1}.Release|x86.Build.0 = Release|Any CPU
+		{8B6BEE0F-F9A5-4E8F-B2DF-56CF504AA272}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8B6BEE0F-F9A5-4E8F-B2DF-56CF504AA272}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8B6BEE0F-F9A5-4E8F-B2DF-56CF504AA272}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{8B6BEE0F-F9A5-4E8F-B2DF-56CF504AA272}.Debug|x86.Build.0 = Debug|Any CPU
+		{8B6BEE0F-F9A5-4E8F-B2DF-56CF504AA272}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8B6BEE0F-F9A5-4E8F-B2DF-56CF504AA272}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8B6BEE0F-F9A5-4E8F-B2DF-56CF504AA272}.Release|x86.ActiveCfg = Release|Any CPU
+		{8B6BEE0F-F9A5-4E8F-B2DF-56CF504AA272}.Release|x86.Build.0 = Release|Any CPU
+		{F77D3A11-DABB-4C35-B388-73E6F50628AC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F77D3A11-DABB-4C35-B388-73E6F50628AC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F77D3A11-DABB-4C35-B388-73E6F50628AC}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{F77D3A11-DABB-4C35-B388-73E6F50628AC}.Debug|x86.Build.0 = Debug|Any CPU
+		{F77D3A11-DABB-4C35-B388-73E6F50628AC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F77D3A11-DABB-4C35-B388-73E6F50628AC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F77D3A11-DABB-4C35-B388-73E6F50628AC}.Release|x86.ActiveCfg = Release|Any CPU
+		{F77D3A11-DABB-4C35-B388-73E6F50628AC}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{AA86F904-90E5-407F-87B6-688ED8A2FF85} = {4186BBA3-FBD8-408B-AF5A-2848A4073DD1}
-		{3B82BB93-FD1B-48E9-936A-FECE89810B16} = {4186BBA3-FBD8-408B-AF5A-2848A4073DD1}
+		{8B6BEE0F-F9A5-4E8F-B2DF-56CF504AA272} = {4186BBA3-FBD8-408B-AF5A-2848A4073DD1}
+		{F77D3A11-DABB-4C35-B388-73E6F50628AC} = {4186BBA3-FBD8-408B-AF5A-2848A4073DD1}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {C570FD29-9F0C-417A-BB07-2135E30E76F3}
